+++ backend/greenwebanalyzer/views.py
@@ -55,7 +55,7 @@ def setupRoutes(app, limiter):
             app.logger.error("Invalid URL was given: %s", url)
             abort(400, description="No valid URL given.")
 
-        r = Report(url)
+        r = Report(url, app=app)
         report = r.create_report()
 
         response = make_response(jsonify(report), 201)
@@ -76,6 +76,8 @@ def setupRoutes(app, limiter):
         try:
             url = json.loads(response.data)['url']
         except JSONDecodeError:
+            url = None
+        except KeyError:
             url = None
 
         app.logger.info({